switch appliance_id to positive small integer field as a fix

+++ backend/apps/home/models.py
@@ -48,7 +48,7 @@ class Appliance(PolymorphicModel):
     """
 
     name = models.CharField(verbose_name="Name", max_length=50)
-    appliance_id = models.IntegerField(verbose_name="Appliance ID")
+    appliance_id = models.PositiveSmallIntegerField(verbose_name="Appliance ID")
     room = models.ForeignKey(
         Room, verbose_name="Room", related_name="appliances", on_delete=models.CASCADE
     )
